compute pressure: Adding supportedSourceTypes to API
To match existing observers API,
supportedSourceTypes read-only property has been added
to the Compute Pressure interface.

https://wicg.github.io/compute-pressure/#the-supportedsources-attribute

It returns an array of SourceType values supported but the user agent.
As the list of supported sources may vary per browser
and will be evolving, this property allows web developers
to check which are available.

This patch doesn't affect the functionality of compute pressure.

Change-Id: Ie1822dd90dcc4027624fc56d80424e6ea5484da1
Bug: 1292466